refactor: handle pools correctly
